+++ lib/src/source_visitor.dart
@@ -9,6 +9,8 @@ import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/source/line_info.dart';
 // ignore: implementation_imports
 import 'package:analyzer/src/clients/dart_style/rewrite_cascade.dart';
+// ignore: implementation_imports
+import 'package:analyzer/src/dart/ast/ast.dart' as src_ast;
 
 import 'argument_list_visitor.dart';
 import 'ast_extensions.dart';
@@ -1813,7 +1815,7 @@ class SourceVisitor extends ThrowingAstVisitor {
     var hasInnerControlFlow = false;
     for (var element in ifElements) {
       _visitIfCondition(element.ifKeyword, element.leftParenthesis,
-          element.condition, element.caseClause, element.rightParenthesis);
+          element.expression, element.caseClause, element.rightParenthesis);
 
       visitChild(element, element.thenElement);
       if (element.thenElement.isControlFlowElement) {
@@ -1856,7 +1858,7 @@ class SourceVisitor extends ThrowingAstVisitor {
 
   @override
   void visitIfStatement(IfStatement node) {
-    _visitIfCondition(node.ifKeyword, node.leftParenthesis, node.condition,
+    _visitIfCondition(node.ifKeyword, node.leftParenthesis, node.expression,
         node.caseClause, node.rightParenthesis);
 
     void visitClause(Statement clause) {
@@ -2252,7 +2254,22 @@ class SourceVisitor extends ThrowingAstVisitor {
 
   @override
   void visitNamedType(NamedType node) {
-    visit(node.name);
+    final importPrefix = node.importPrefix;
+    if (importPrefix != null) {
+      // TODO(scheglov) This is wrong, we should not create nodes here.
+      // But I don't see how to change CallChainVisitor() to work with tokens.
+      // So, I will leave it to the maintainers to figure this out.
+      visit(
+        src_ast.PrefixedIdentifierImpl(
+          prefix: src_ast.SimpleIdentifierImpl(importPrefix.name),
+          period: importPrefix.period,
+          identifier: src_ast.SimpleIdentifierImpl(node.name2),
+        ),
+      );
+    } else {
+      token(node.name2);
+    }
+
     visit(node.typeArguments);
     token(node.question);
   }